Change to unsignedInteger

+++ database/migrations/2016_04_19_112407_create_store_notification_requests_table.php
@@ -15,8 +15,8 @@ class CreateStoreNotificationRequestsTable extends Migration
         Schema::connection('mysql-store')->create('notification_requests', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
-            $table->mediumInteger('user_id');
-            $table->integer('product_id');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('product_id');
 
             $table->unique(['user_id', 'product_id']);
         });
